+++ lib/ua_inspector/parser/device.ex
@@ -8,8 +8,11 @@ defmodule UAInspector.Parser.Device do
   alias UAInspector.Parser.VendorFragment
   alias UAInspector.Result
   alias UAInspector.Util
+  alias UAInspector.Util.Fragment
 
   @behaviour UAInspector.Parser.Behaviour
+
+  @frozen_android Regex.compile!("Android 10[.\d]*; K(?: Build/|[;)])", [:caseless])
 
   @hbbtv Util.build_regex("HbbTV/([1-9]{1}(?:\.[0-9]{1}){1,2})")
   @notebook Util.build_regex("FBMD/")
@@ -17,18 +20,9 @@ defmodule UAInspector.Parser.Device do
 
   @impl UAInspector.Parser.Behaviour
   def parse(ua, client_hints) do
-    agent_result =
-      cond do
-        Regex.match?(@hbbtv, ua) -> parse_hbbtv(ua)
-        Regex.match?(@shelltv, ua) -> parse_shelltv(ua)
-        Regex.match?(@notebook, ua) -> parse_notebook(ua)
-        true -> parse_regular(ua)
-      end
-      |> maybe_parse_vendor(ua, client_hints)
-
-    hints_result = parse_hints(client_hints)
-
-    merge_results(hints_result, agent_result)
+    client_hints
+    |> parse_hints()
+    |> parse_device(client_hints, ua)
   end
 
   @doc """
@@ -48,6 +42,61 @@ defmodule UAInspector.Parser.Device do
   @spec shelltv?(String.t()) :: boolean
   def shelltv?(ua), do: Regex.match?(@shelltv, ua)
 
+  defp parse_device(%{model: :unknown} = hints_result, client_hints, ua) do
+    if Fragment.desktop?(ua) do
+      hints_result
+    else
+      parse_device_details(hints_result, client_hints, ua)
+    end
+  end
+
+  defp parse_device(
+         %{model: model} = hints_result,
+         %{platform_version: platform_version} = client_hints,
+         ua
+       ) do
+    ua =
+      if Regex.match?(@frozen_android, ua) do
+        os_version =
+          case platform_version do
+            :unknown -> "10"
+            "" -> "10"
+            _ -> platform_version
+          end
+
+        Regex.replace(
+          ~r/(Android 10[.\d]*; K)/,
+          ua,
+          "Android #{os_version}; #{model}"
+        )
+      else
+        ua
+      end
+
+    parse_device_details(hints_result, client_hints, ua)
+  end
+
+  defp parse_device(hints_result, client_hints, ua),
+    do: parse_device_details(hints_result, client_hints, ua)
+
+  defp parse_device_details(hints_result, client_hints, ua) do
+    agent_result =
+      cond do
+        Regex.match?(@hbbtv, ua) -> parse_hbbtv(ua)
+        Regex.match?(@shelltv, ua) -> parse_shelltv(ua)
+        Regex.match?(@notebook, ua) -> parse_notebook(ua)
+        true -> parse_regular(ua)
+      end
+      |> maybe_parse_vendor(ua, client_hints)
+
+    merge_results(hints_result, agent_result)
+  end
+
+  defp merge_results(%{} = hints_result, %{brand: :unknown, model: :unknown, type: :unknown}),
+    do: hints_result
+
+  defp merge_results(_, agent_result), do: agent_result
+
   defp do_parse(_, []), do: :unknown
 
   defp do_parse(ua, [{regex, {_, models, _, _} = device_result} | database]) do
@@ -63,11 +112,6 @@ defmodule UAInspector.Parser.Device do
   end
 
   defp maybe_parse_vendor(device, _, _), do: device
-
-  defp merge_results(%{} = hints_result, %{brand: :unknown, model: :unknown, type: :unknown}),
-    do: hints_result
-
-  defp merge_results(_, agent_result), do: agent_result
 
   defp parse_hints(%{model: model}) when is_binary(model),
     do: %Result.Device{model: Util.maybe_unknown(model)}