+++ lib/ua_inspector/util.ex
@@ -16,10 +16,12 @@ defmodule UAInspector.Util do
     )
   end
 
-  def build_engine_regex("Blink"), do: build_engine_regex("Chrome")
+  def build_engine_regex("Blink"), do: build_engine_regex("Chrome|Cronet")
 
   def build_engine_regex(name) do
-    Regex.compile!(name <> ~S"\s*\/?\s*((?(?=\d+\.\d)\d+[.\d]*|\d{1,7}(?=(?:\D|$))))", [:caseless])
+    Regex.compile!("(?:" <> name <> ~S")\s*\/?\s*((?(?=\d+\.\d)\d+[.\d]*|\d{1,7}(?=(?:\D|$))))", [
+      :caseless
+    ])
   end
 
   @doc """
@@ -31,6 +33,14 @@ defmodule UAInspector.Util do
   @spec build_regex(regex :: String.t()) :: Regex.t()
   def build_regex(regex) do
     Regex.compile!("(?:^|[^A-Z0-9\-_]|[^A-Z0-9\-]_|sprd-|MZ-)(?:" <> regex <> ")", [:caseless])
+  end
+
+  @doc """
+  Build a generic matching regex.
+  """
+  @spec build_base_regex(regex :: String.t()) :: Regex.t()
+  def build_base_regex(regex) do
+    Regex.compile!("(?:^|[^A-Z_-])(?:" <> regex <> ")", [:caseless])
   end
 
   @doc """